outside/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-google-genai
bcrypt
mysql-connector-python
python-dotenv
pandas
faiss-cpu
